backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
spotipy==2.23.0
requests==2.31.0
redis==5.0.1
pydantic==2.5.0
PyJWT==2.8.0
cryptography==41.0.7
ytmusicapi==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
redis==5.0.1